+++ exsampleFiles/ConfigManager.py
@@ -1,3 +1,6 @@
+from GabesPythonToolBox.Utility.Debug import Debug
+#Debug.add_group('LIB', True) #Uncoment for the debug mesages
+
 import GabesPythonToolBox.Utility.ConfigManager as CM
 
 # Initialize the ConfigManager with the file path
@@ -8,8 +11,8 @@ config = config_manager()
 print("Loaded Configuration:")
 print(config)
 
-# Update the player's name to "Sylvanas"
-config_manager.update("player.name", "Sylvanas")
+# Update the player's name to "Opydian"
+config_manager.update("player.name", "Opydian")
 print("\nUpdated Player Name:", config_manager()["player"]["name"])
 
 # Change the game difficulty to "hard"
@@ -18,7 +21,7 @@ print("\nUpdated Game Difficulty:", config_manager()["game_settings"]["difficult
 
 # Add a new item to the player's inventory
 current_inventory = config_manager()["player"]["inventory"]
-current_inventory.append("magic ring")
+current_inventory.append("Dragon Scale")
 config_manager.update("player.inventory", current_inventory)
 print("\nUpdated Inventory:", config_manager()["player"]["inventory"])
 
@@ -27,7 +30,7 @@ config_manager.update("quests.active[0].progress", 80)
 print("\nUpdated Quest Progress (Defeat the Dragon):", config_manager()["quests"]["active"][0]["progress"])
 
 # Save the updated configuration back to the Newfile
-config_manager.saveNew("exsampleFiles\ConfigManager2.json")
+config_manager.save("exsampleFiles\ConfigManager2.json")
 print("\nnew file saved:")
 print(config_manager())
 
@@ -37,5 +40,5 @@ print("\nConfiguration Reset:")
 print(config_manager())
 
 # Save the updated configuration back to the file
-config_manager.save()
+config_manager.save("exsampleFiles\ConfigManager.json")
 print("\nConfiguration has been saved back to the file.")