+++ ChangeLog.md
@@ -1,5 +1,44 @@
 # Updates
 
+## 1.5.1 Debug update 
+
+### bug Fix
+
+- __init__ file update
+
+### changes
+
+- Added lib debug mesages to entire librery (not exsamples will be added soon)
+
+## 1.5.0 Debug update 
+
+### new things
+
+- Debug Lib
+
+## 1.4.0 UnitConverter update 
+
+### new things
+
+- UnitConverter
+
+### changes
+
+- changed how the interaction with config manager works
+
+## 1.3.0 Collor update 
+
+### new things
+
+- Collor Lib
+
+## 1.2.0 Config Manager update 
+
+### new things
+
+- Config Manager
+
+
 ## 1.1.1 doc update
 
 ### new things